fixed issues.
