+++ burn-core/src/nn/loss/mod.rs
@@ -1,7 +1,9 @@
+mod binary_cross_entropy;
 mod cross_entropy;
 mod mse;
 mod reduction;
 
+pub use binary_cross_entropy::*;
 pub use cross_entropy::*;
 pub use mse::*;
 pub use reduction::*;
